+++ engine/src/flutter/impeller/entity/contents/text_contents.cc
@@ -77,6 +77,22 @@ void TextContents::SetTextProperties(Color color,
   }
 }
 
+namespace {
+Scalar AttractToOne(Scalar x) {
+  // Epsilon was decided by looking at the floating point inaccuracies in
+  // the ScaledK test.
+  const Scalar epsilon = 0.005f;
+  if (std::abs(x - 1.f) < epsilon) {
+    return 1.f;
+  }
+  if (std::abs(x + 1.f) < epsilon) {
+    return -1.f;
+  }
+  return x;
+}
+
+}  // namespace
+
 void TextContents::ComputeVertexData(
     VS::PerVertexData* vtx_contents,
     const std::shared_ptr<TextFrame>& frame,
@@ -165,7 +181,8 @@ void TextContents::ComputeVertexData(
         atlas_glyph_bounds = maybe_atlas_glyph_bounds.value().atlas_bounds;
       }
 
-      Rect scaled_bounds = glyph_bounds.Scale(1.0 / rounded_scale);
+      Scalar inverted_rounded_scale = 1.f / rounded_scale;
+      Rect scaled_bounds = glyph_bounds.Scale(inverted_rounded_scale);
       // For each glyph, we compute two rectangles. One for the vertex
       // positions and one for the texture coordinates (UVs). The atlas
       // glyph bounds are used to compute UVs in cases where the
@@ -175,8 +192,14 @@ void TextContents::ComputeVertexData(
       Point uv_size = SizeToPoint(atlas_glyph_bounds.GetSize()) / atlas_size;
 
       Matrix unscaled_basis =
-          Matrix::MakeScale({basis_transform.m[0] > 0 ? 1.f : -1.f,
-                             basis_transform.m[5] > 0 ? 1.f : -1.f, 1.f});
+          basis_transform * Matrix::MakeScale({inverted_rounded_scale,
+                                               inverted_rounded_scale, 1});
+
+      // In typical scales < 48x these values should be -1 or 1. We round to
+      // those to avoid inaccuracies.
+      unscaled_basis.m[0] = AttractToOne(unscaled_basis.m[0]);
+      unscaled_basis.m[5] = AttractToOne(unscaled_basis.m[5]);
+
       Point unrounded_glyph_position =
           // This is for RTL text.
           unscaled_basis * glyph_bounds.GetLeftTop() +